arm64/s2mpu: Use new pkvm_iommu_register function signature
pkvm_iommu_register was changed in ACK to accept an extra argument
"flags", so device specfic flags can be passed to the hypervisor
driver as presence of sync devices.
This patch changes pkvm_iommu_register calls to have the new arg.

The flags will be used later when sync devices are supported.

Bug: 257544191
Change-Id: Ibbd3d7a79e051e8fc8765e683399644cae4965f6
Signed-off-by: Mostafa Saleh <smostafa@google.com>