google/dts: gs201: update pinctrl DT header
The upstream commit 3d501682329a ("arm64: dts: exynos: use local header
for pinctrl register values") deprecated the samsung pinctrl.h
dt-binding in favor of using a headering directly in the dts folder. The
motive was that adding raw register values to DT bindings does not fit
the purpose of bindings.  So update our DTS to use the exynos-pinctrl.h
header and also move the gs101 pinctrl header over to the dts folder as
well and rename it gs101-pinctrl.h.

Change-Id: Ib30e8e9dc3ea76140086f78f578b3417d6f5e3ef
Signed-off-by: Samuel Gosselin <sgosselin@google.com>