s2mpg13_spmic: Switch to new OF thermal API
The thermal OF code has a new API allowing to migrate the OF
initialization to a simpler approach. The ops are no longer device
tree specific and are the generic ones provided by the core code.

Convert the ops to the thermal_zone_device_ops format and use the new
API to register the thermal zone with these generic ops.

Change-Id: Id2d0f4b52c9a108aebb5e9bf05064f33eb107cc1
Signed-off-by: Samuel Gosselin <sgosselin@google.com>