vh: sched: Split DVFS headroom calculations from mapping to freq
With uclamp in the picture, the rules are getting complex and the
mapping function is now overloaded.

We need to prevent creating an additional headroom (aka
margin/boost) that can either break uclamp_max contract, or apply an
additional boost on top of uclamp_min boost.

The right to place to apply the headroom is when calculating the util
for the rq. This should give us the right propagation automatically.

Remove cpu arg from map_util_freq_pixel_mod() as it's no longer
required. And max_cpu arg from em_cpu_energy_pixel_mod(). Effectively
partially undoing previous changes to fix similar problem.

Bug: 278091167
Signed-off-by: Qais Yousef <qyousef@google.com>
Change-Id: Idb2f971e4a59a6a67c97cd6706fffaa508839b12