bcl: Split get/clear max77759 IRQ into 2 functions
Over current reboot reason are classified as Under Voltage reboot reason
because BATOILO IRQ was clear as soon as it is triggered.
MAX77759 classified over current reboot reason if there is an unclear
BATOILO IRQ.  Therefore in this change, we delay the clearing of the IRQ
to after when device starts throttling.

Bug: 277976680
Change-Id: I26c75e823602ac0f59791957d6386e6216a4c4b7
Signed-off-by: George Lee <geolee@google.com>