media: mfc: manages the count for display DRC
When the VP9 interframe resolution change stream is
decoded to NAL_Q mode, there is a problem that
the display DRC flag is cleared when the display DRC
occurs in two consecutive frames.

1. #N frame: NAL_Q_DONE with display DRC "disp_res_change = 1"
2. #N frame: wait G_FMT
3. #N+1 frame: NAL_Q_DONE with display DRC "disp_res_change = 1"
4. #N frame: g_fmt "disp_res_change = 0"
5. #N+1 frame: g_fmt "DPB size updated because there is no disp_res_change"
This can cause error No.57(INSUFFICIENT_DPB_SIZE).

So this patch is managed as count so as not to clear
disp_res_change flag for frames that need g_fmt.

Bug: 269568991
Signed-off-by: AyoungSim <a.sim@samsung.com>
Change-Id: I503c8823cb054d78a5b6b43b52e92ab014374230
(cherry picked from commit 1c4fa489cd655fe988be0036d50bc7f21bcfbde9)
Signed-off-by: Will McVicker <willmcvicker@google.com>