pcie: exynos: add reg lock when the phy soft reset
If PCIe stack try to access the pci config register
for entering the suspend, and the exynos pci is processing
the exynos_pcie_poweroff function at the same time,
accessing the pci config during the phy reset can be happened.
and it cause problems. To prevent this problem, add the lock during reset.
also, register the bridge ops by deleting the condition.

Bug: 278939245

Change-Id: I5ef9108b4a646e55421c326751951ccb4a1bc2c1
Signed-off-by: Kyounghye Yun <k-hye.yun@samsung.com>