PIXEL: ufs: add kibble trigger support for power events
This CL adds a feature to the UFS driver to update Kibble triggers on
relevant power events. The UFS driver will make the following
associations with UFS regions:

* ufs_active_idle - Used to track the UFS device being in either Active
or Idle state; i.e. powered on but outside of hibernate.
* ufs_power_on - Used to track the UFS controller/device being powered
on/off. UFS is only powered off during system suspend - for which the
trigger can be used to track.
* ufs_io_outstanding - Used to track any outstanding I/O on the UFS
device. The trigger will be off when there are no I/O commands
outstanding.

The feature is disabled by default, and is enabled by writing 1 to the
newly added sysfs entry under pixel/power_event_mode.

Bug: 288171470
Change-Id: Ia452cfcc7599bdd1c04e8146c840a5a4904217a8
Signed-off-by: Konstantin Vyshetsky <vkon@google.com>