kleaf: make the GKI directory selectable
This adds a gki_kernel_dir flag that can be used to select between the
aosp or aosp-staging kernel. By default it is set to use the "aosp"
kernel directory. To use the aosp-staging directory, you can build with
--config=aosp_staging.

This patch also updates the build config fragment logic to be able to
apply the debug build configs to the AOSP GKI kernel. This is done by
setting the --gki_build_config_fragment=... flag in the device.bazelrc
files along side the debug_fragment flags.

Bug: 283159637
Change-Id: If1c60f3b8a484c94586bdab75ada59c420cf14d6
Signed-off-by: Will McVicker <willmcvicker@google.com>