pcie: drivers: Change Wifi RefClk off time to 4us
When CLKREQ is deasserted, RefClk turns off within 4us for modem but
takes 75us for wifi. This could lead to instabilities in the pcie link
for short CLKREQ pulses where the EP deasserts and then reasserts the
clock within a short time. In this case when the CLKREQ is reasserted
the EP will see the RefClk for a short amount of time, but then 74us
later the RefClk dissapears and then reappears. Reducing this time and
making it match the modem reduces the window for any EP transactions
within this time period.

Bug: 279486159
Change-Id: Ib842cfcd0a21bea3e03671434b3fd186388ef8f9
Signed-off-by: Sajid Dalvi <sdalvi@google.com>
(cherry picked from commit 7b56db38729c511ebebe88bca66fe3b1524fbe10)