vh: sched: Fix handling of root group and autogroup uclamp_min aggregation
Root group and autogroup uclamp_min is set as a limit as documented in:

	https://kernel.org/doc/html/latest/scheduler/sched-util-clamp.html?highlight=utilization+clamp#sched-util-clamp-min

Which means we should not use their uclamp_min as a minimum value, but
rather as an upper bound of what the user can request.

The handling of setting this upper limit is in uclamp_eff_get() where if
the aggregated value is greater than root group uclamp_min, then we cap
it to the root group value.

Fix our handling to match the upstream definition.

Bug: 281598730
Signed-off-by: Qais Yousef <qyousef@google.com>
Change-Id: I6beecaddc818d823afa641764e21c5e317cb3808
(cherry picked from commit 53a55decc47ce8e88abbc0bec382e71a3869cc00)
Signed-off-by: Will McVicker <willmcvicker@google.com>