Add an API allowing multiple return values from SMC invocations
The existing exynos_smc API takes 3 input parameters that are passed
along with the SMC command code, but returns only one value, which is
typically the operation result or the error code. In some cases we need
to return more than just status/error code, for example when reading a
register value via SMC. This new API allows returning up to 3
additional values that are distinct from the operation error code.

Bug: 238380992
Test: read secure CRCs via sysfs
Change-Id: I9e6f7491db9f97d4cddcc12eb2a3d81461950c52
Signed-off-by: Sergey Volk <servolk@google.com>