tcpci_max77759: enable/disable dp pull up regulator through sysfs
When setting the sbu state through the max77759's sbu_pullup sysfs
node, enable or disable the DisplayPort pull up regulator and set
the voltage to the dp specified level or to the specsheet defined
min/max respectively.

echoing values 0 and 1 disable the regulator for normal and reverse
orientations respectively, while values 2 and 3 enable the regulator.

Test: manual test on device
Bug: 283908925
Change-Id: Ibf8c2245fa03f4736d5ec31b71fba12d17cf6c83
Signed-off-by: RD Babiera <rdbabiera@google.com>