tcpci_max77759: Defer checking for contaminant when throttled
Defer checking for contaminant and compliance warnings when USB port is
throttled due to overheat. This change introduces USB_THROTTLE_VOTABLE
which is set when the port is disabled by the usbc-port cooling device
and cleared when the port is re-enabled. Contaminant and missing-rp
compliance warnings are deferred when the USB_THROTTLE_VOTABLE is set.

Bug: 287333701
Change-Id: I4c69ef0914449cb7d9fcc2d55be60ac13798e291
Signed-off-by: Badhri Jagan Sridharan <badhri@google.com>
(Cherry-picked from commit a151619d9e48137f3261d08503bb6be3db97b6d7)