pcie: exynos: Check HSI1 PMU reset
This fixes the ITMON while reading 0x1213C700.
Analyzing the scan dump of the failure, HSI1 is in reset. This should
not be the case, but can explain why we are seeing the ITMON. The
change here checks the HSI1 related reset status and if it is in reset,
then forces it out of reset.

Bug: 287328303
Change-Id: I543f5f200bb3b664c83a8fd9c590332185c7267e
Signed-off-by: Kyounghye Yun <k-hye.yun@samsung.com>
Signed-off-by: Sajid Dalvi <sdalvi@google.com>