pcie: exynos: improve the TX stability
This fixes the PCIe link down issues seen on the WIFI interface.
The TX FIFO acts as a buffer between the digital and analog logic.
Link TX data -> PMA TX FIFO -> PMA TX DP/DN
By setting 0x1350 to high, the analog (PLL) clock to TX FIFO can
be gated, until the PLL becomes stable (a few cycles after PLL
lock is done). The default setting is low, which supplies the PLL
lock to TX FIFO as soon as the PLL lock is done, which could cause
instabilities in the TX data, which could result in PCIe link down.

Bug: 286689794
Change-Id: Ic6de35fd4f675fd6cccbd1aad748c17494d08a86
Signed-off-by: Kyounghye Yun <k-hye.yun@samsung.com>
Signed-off-by: Sajid Dalvi <sdalvi@google.com>