pcie: Cleanup root port config space save / restore
The root port config space registers should be saved at the beginning
of poweroff() to retain any modifcations.  They should be restored at
the beginning of poweron() so that any subsequent modifications aren't
overwritten by the restore.

As part of PCIe RC bridge enumeration, the upstream pcieport driver
performs a pci_save_state().  These values should be stored at probe
so they are available on the first poweron() call.  To do this, the
detection and saving of pci_dev moved from poweron to probe.

Bug: 291068808
Signed-off-by: Doug Zobel <zobel@google.com>
Change-Id: Idbacfc8f650abfc4030c468888fe0ea827a33349