phy-exynos-usbdrd: Power manage HSI0-PD S2MPU without RPM
In GS201, the behaviour was to eagerly resume/suspend s2mpu from post
ON and pre OFF hooks, without RPM.

In Zuma, to avoid symbol dependency between exynos-pd (which depends
on this driver) calls to S2MPU driver were removed, and the suspend
/resume was done through kernel run time PM through put/get sync.

However, it turned out this doesn't work during system suspend, as
the kernel disable rpm for devices which prevents the suspend
notification to reach the S2MPU driver, so the hypervisor still
thinks the S2MPU is on while it is not, causing the crash upon MMIO
access.

This is done from: pm_suspend => enter_state =>
suspend_devices_and_enter => dpm_suspend_start => dpm_prepare =>
device_prepare => pm_runtime_get_noresume

To fix this, we use a similar mechanism to Pixel-7, to directly
notify the hypervisor.

To avoid symbol dependency, the driver calls to KVM IOMMU directly,
which is safe.

Bug: 292146325
Change-Id: I6d4b0024422caa58b6cef78d8ac43e6b367e3800
Signed-off-by: Mostafa Saleh <smostafa@google.com>