update_symbol_list.sh: Hardcode the GKI_BRANCH
We need to be pushing symbol updates to the main GKI branch, e.g.
android14-5.15 and not release branches. So up the GKI_BRANCH define to
use that directly. To update the release branch, you can cherry-pick the
main patch and file a respin bug.

This fixes issues updating the symbol list now that aosp/ is on
a release branch.

Bug: 286989839
Change-Id: Ibe083cdd3d53268a058125194ce46d8387900aab
Signed-off-by: Will McVicker <willmcvicker@google.com>