add arm64/dts: faceauth_eval_heap and faeval.
This is a dynamic (optional) heap for faceauth for eval.
Currently, faceauth_rawimg_heap is used. This is the first step to move
part of the allocations from faceauth_rawimg_heap to faceauth_eval_heap.
The new, faceauth_eval_heap, will be used as an optional heap (reserved
only when used), while faceauth_rawimg_heap is always reserved and
required.

Change-Id: I44637cb368f711dec908a00b439e8b8a64e86337
Signed-off-by: Saeed Alrahma <alrahma@google.com>
Bug: 283294203