dwc3-exynos: Fix up condition for force enabling device mode
While in unplugged state its valid for extcon to return 0 for
both EXTCON_USB and EXTCON_USB_HOST. EXTCON APIS also return
errnos to propogate error condition. Given that the intention
is to only force enable device mode when extcon property is
not set, its safer to directly check for it.

Bug: 300138471
Test: usb bootup and plug/unplug test on dev board
Change-Id: Id4b9e5f6307cd0a656c00f4f359944034d2fcaad
Signed-off-by: Badhri Jagan Sridharan <badhri@google.com>