soc/google: make hardlockup watchdog configurable from command line
It was previously only possible to disable the EHLD and the kernel
builtin softlockup detector from the command line. With this change,
make it possible to prevent the hardlockup watchdog from panicking
the kernel via:

hardlockup_watchdog.hardlockup_panic=0

With this it is now possible to disable all panics triggered by
watchdogs and support intrusive kernel debugging.

Bug: 303691900
Change-Id: Ie9a24a72e225143f58bcdd955320a25151204b15
Signed-off-by: Peter Collingbourne <pcc@google.com>