arm64/dts: zumapro: debug: Add a compatible string
Add a compatible string to the dpm overlays so that you can easily
verify if they are in the dtbo.img. When included, the overlay
compatible string is included in the dt_table_entry. For example, if you
include the dpm-user overlay with this patch, you'll see this in the
output:

  $ mkdtimg dump dtbo.img:
  ...
  dt_table_entry[70]:
               dt_size = 4096
             dt_offset = 8047871
                    id = 00000000
                   rev = 00000000
             custom[0] = 00000000
             custom[1] = 00000000
             custom[2] = 00000000
             custom[3] = 00000000
             (FDT)size = 4096
       (FDT)compatible = google,ZUMAPRO DPM-user

Bug: 296637949
Change-Id: I89972500521ff546fb2d8e429031935645cf0502
Signed-off-by: Will McVicker <willmcvicker@google.com>