usb: dwc3: place the might-sleep function out of atomic context
This patch moves the might-sleep function of
dwc3_exynos_clk_disable_unprepare out of atomic context.

The dwc3_exynos_clk_disable_unprepare function calls the clk_unprepare
function, which should not be called from within atomic context. This
is because stopping a clock requires time for the hardware to be
unprepared, which may cause the process to sleep during this time. If
the clk_unprepare function is called from within atomic context, no one
can wake the process up once it has slept.

Bug: 289587519
Test: dwc3 functionality including MTP, tethering function, android backup, suspend/resume and replug tests. (test request b/308377129)
Change-Id: I6bc4cc480d99f6feaed7a564c2f79a5cb0e2fadc
Signed-off-by: Kuen-Han Tsai <khtsai@google.com>