media: mfc: prevent idle suspend when MFC is attempting a H/W operation
In certain corner cases, the idle timer may trigger idle mode when MFC
is attempting a hardware operation. To avoid this timing issue, we have
added a hw_run_cnt to ensure that the idle timer will not trigger idle
mode for at least 1.5 seconds when MFC is attempting a hardware
operation.

Bug: 309545436
Test: SST, secure/non-secure playback pause/resume,
foreground/background switch.
Change-Id: I4f79757bfbbfc37f629cf35a0067e7352d8ff690
Signed-off-by: wenchangliu <wenchangliu@google.com>