soc: google: gsa: create GSA Char Device
Currently there is no extensible way to communicate with GSA
from Android. We need this capability at least for dynamic app
loading on GSA, and it is likely that there will be additional
tasks that require communicating with GSA from Android.

This patch is the interface for the character device ioctl that
is being implemented to handle app loading. Currently the ioctl
will just return an error indicating it is not implemented yet.

This patch includes gsa.h, a userspace API for communicating
with the ioctl.

Bug: 293333685
Test: Module loads and cdev is present. Sending correct ioctl
results in expected error logged.
Change-Id: Icea822ce63f8a33a2df8e258aa72e7434ffe2e94
Signed-off-by: Anthony LaCross <lacross@google.com>