pcie: exynos: Move dw_pcie_host_init out of probe
dw_pcie_host_init has a one second delay waiting for the end point
driver to enumerate. In our case, the end point drivers are initialized
in their probe/init functions which come later in the sequence.
This change moves the dw_pcie_host_init and the associated code out of
the probe and in the context of the endpoint driver init.
This saves 1 sec per pcie interface during the initial probe sequence.

Bug: 315052790

Change-Id: Ifd758d4b1213311f9f831ec4d8f5740bd3e9a96d
Signed-off-by: Sajid Dalvi <sdalvi@google.com>