vh: sched: Ensure dequeue/enqueue functions start working together
We have a race where enqueue being registered before dequeue introducing
a race where a task can be enqueued AND dequeued between the two
registration calls. This might lead to unpleasant events later like
potential list corruption if a node was added but not removed.

This only fixes the race of enqueue called without dequeue. But it
doesn't fix the race that a dequeue is called without enqueue which can
happen a task was already enqueued before our hook is registered.

There's not much we can do about that except making sure our dequeue
path is safe to such events with adequate protections.

Bug: 296116217
Signed-off-by: Qais Yousef <qyousef@google.com>
Change-Id: Iefc5b26652b554bf50d314cd68df5e59120379da