vh: sched; Implement pseudo per-task iowait boost
We still rely on the per-cpu boost mechanism, but we add our own
tracking for iowait boost for each task so that it can help us in task
placement to place it on bigger cores; and on migration to help kick
start iowait boost on the new CPU.

Allow controlling the maximum achievable per task iowait boost as the
current boost behavior can be too aggressive potentially. And the
percpu iowait boost value too.

Clean up the code where a post_util_init_scale_show() function was
misplaced and SCHED_CAPACITY_SCALE was open coded to 1024 + whitespace
cleanups.

Bug: 318366482
Signed-off-by: Qais Yousef <qyousef@google.com>
Change-Id: I14f265272c5c472e39a5798761f9e8c47da57b80