vh: sched: Make fits_capacity() automatic
Fits capacity relied on magic margin value to decide when a task does
not fit a CPU. Make the definition more deterministic by making sure
that undoing the invariance and taking TICK_USEC into account.

capacity and frequency invariance means that the task will have to run
longer on smaller cores to reach a specific utilization value. To ensure
the a task will move smoothly between capacities and not perceive a big
delay, we need to undo invariance so that the time it takes the task to
reach capacity_of(small_core) is the same across all CPUs.

And to handle the fact that misfit migration can be delayed by TICK
triggering the misfit migration as a worst case scenario, we need to
give ourselves enough headroom to cater for this worst case scenario.

The updated logic will mean that we should move faster from little cores
and slower from mid cores and not trigger (take much longer) misfit on
biggest cores. Which means overutilized will now take longer to trigger
on big core too.

group_overutilized() still uses the old cpu_overutilized() function.
Only util_fits_cpu() was updated.

Bug: 315748906
Signed-off-by: Qais Yousef <qyousef@google.com>
Change-Id: I2ae776b469ce3b6e41bda26a364c50249ba3d2d4