soc: google: gsa: Add GSA AppLoader IOCTL functionality
GSA has a mailbox command to dynamically load apps at
run-time instead of including them at build-time. At
the moment this can only be done via fastboot, not
userspace. We want to do this during android boot when
the init scripts run so Tuscany can be loaded to GSA.
Right now it is loaded to TZ.

This patch adds ioctl functionality for app loading from
userspace.

Bug:293333685
Test: Loaded sample app using this ioctl and checked GSA
logs to see that it ran as expected.
Signed-off-by: Anthony LaCross <lacross@google.com>

Change-Id: I3ce803d0386a5a0598414d1fe2d77fac6dfd5e84