misc: sbb-mux: Add cp2ap_wake_isr and pcie_link_state SBB signals
cp2ap_wake_isr: triggered when the modem isr for cp2ap_wake is asserted
or de-asserted by the modem chip.
pcie_link_up: tracks the PCIe link state. It will be high when the link
enters L0 and will be low when the link transitions to L2. Because L1SS
transitions are handled by the hardware, the software will not be able
to track these power states. PCIe power can be used to track these
intermediate states.

Bug: 323386158
Change-Id: I1f5cbd48389aca3d2a42a35597e08bfc55b59636
Signed-off-by: Sajid Dalvi <sdalvi@google.com>