pcie: exynos: Add SBB debug signals to track pcie link state
SBB_SIG_MODEM_PCIE_LINK_STATE will be used to track the pcie link state.
It will be high when the link enters L0 and will be low when it
transitions to L2. Because L1SS transitions are handled by hardware,
they will need to be tracked using PCIe power.

Create a sbb_debug sysfs to enable and disable this functionality.

Bug: 323386158
Change-Id: I00850d660904bf38ebc93ff4db7815b27319af0c
Signed-off-by: Sajid Dalvi <sdalvi@google.com>