cpif: Add SBB debug signals to trigger on cp2ap_wake ISR
SBB_SIG_MODEM_CP2AP_WAKE_ISR will be triggered when the
ap_wakeup_handler() is entered. This will be used to determine the time
between the modem firing the cp2ap_wakeup interrupt and this ISR being
entered. The timing can be used to ensure that there hasn't been any
regression in processing the interrupt and other metric gathering use
cases.

Create a sbb_debug sysfs to enable and disable this functionality.

Bug: 323386158
Change-Id: I84df6382364c18d977ecb7570ced18666e2188c3
Signed-off-by: Sajid Dalvi <sdalvi@google.com>