usb: dwc3: prevent gadget binding during role switch
Gadget driver binding would trigger a dwc3 resume, and it might race
with the resume in dwc3_otg_start_host / dwc3_otg_start_gadget.
Due to the hack in dwc3-exynos to overwrite dwc->current_dr_role in
the resume routine, the resume triggered by gadget driver binding is
undesirable as it might not have the right role to do a proper
resume.
To solve the race, acquire the dev mutex of dwc3 gadget when
switching to host / gadget mode. The mutex is to protect the
access to the driver of the device.

Bug: 317947464
Test: rapid role switch 2000+ iterations
Change-Id: I3fb37b4e36fc2df99ad2a3cd60350f8fda9579a4
Signed-off-by: Roy Luo <royluo@google.com>