vh: sched: Ensure to skip inefficient cpufreq
We tag in the EM now that an OPP is inefficient, but we don't tag it in
the cpufreq core that it is inefficient; which means we can still select
it in cpufreq_gover.c::get_next_freq().

Make sure the wrapper map_util_freq_pixel_mod() handles those efficient
OPPs when doing the mapping so all users will see what the driver should
ultimately select.

Bug: 325030968
Signed-off-by: Qais Yousef <qyousef@google.com>
Change-Id: I4d2ebd982645d8da464802a146bc1008e497a02c