vh: sched: Fix calculation of cost and energy
In our custom energy models the capacities are not linear. The math was
taken from upstream code as the EM doesn't contain a capacity field.

The energy cost should be in general:

	energy = (power * util)/cap  ---- (1)

which due to absence of cap in upstream; it should be

	cap = (freq * max_cap)/max_freq

which means

	energy = (power * max_freq/freq) * util/max_cap

which is what we use. Skip all of this and instead store

	cost = power / cap

and update energy calculation to be the original as in (1)

with the caveat that to keep the resolution high enough; scale with
1024.

Bug: 325030968
Signed-off-by: Qais Yousef <qyousef@google.com>
Change-Id: Id24ea9f70b5eb5ec01b3f46687ec0d15b2038cf9