vh: pixel_em: dts: Add default2 pixel_em profile
It is the same as default but we modify the capacities so that our
reference top capacity is 2.5GHz instead of 2.9/3GHz. This should make
the mid and big CPUs look bigger and allow us to avoid use the big core
as often. Especially when combined with auto migration and dvfs
margins.

We simply scale all values by 1021/898 or 1017/828 (new
capacity@2.5GHz/old capacity@2.5GHz) and simply add 1 for all subsequent
frequencies on the big core.

Bug: 326394523
Signed-off-by: Qais Yousef <qyousef@google.com>
Change-Id: Ib0aff15fc4c55c6c4edfb25261b5f81f2ab5a3ef