vh: sched: Implement auto dvfs headroom
auto dvfs headroom will only apply the bare minimum dvfs headroom to
ensure that the system operates as intended (there's enough idle time
for utilization to grow correctly).

The definition of enough headroom relies on the fact that we are limited
by up_rate_limit_us to increase the frequency when needed. Which means
we really only need to have a small increment until the next frequency
update is able to be executed.

There's a new knob to enable the feature. When it is enabled the
hardcoded dvfs headroom values will be ignored. Including tapered dvfs
headroom.

Bug: 315748906
Signed-off-by: Qais Yousef <qyousef@google.com>
Change-Id: Ie74edb3c72f5fb644ef16dffa8a3e6c57d708fd3