vh: sched: cpufreq_gov: Add new tunable to dictate response_time_ms
The new tunable, response_time_ms,  allow us to speed up or slow down
the response time of the policy to meet the perf, power and thermal
characteristic desired by the user/sysadmin. There's no single universal
trade-off that we can apply for all systems even if they use the same
SoC. The form factor of the system, the dominant use case, and in case
of battery powered systems, the size of the battery and presence or
absence of active cooling can play a big role on what would be best to
use.

The new tunable provides sensible defaults, but yet gives the power to
control the response time to the user/sysadmin, if they wish to.

This tunable is applied before we apply the DVFS headroom for CFS+RT
tasks. It is only effective when auto_dvfs_headroom is enabled. It'd be
a NOP otherwise.

Bug: 315748906
Signed-off-by: Qais Yousef <qyousef@google.com>
Change-Id: I27459dd7b3eaebeab4ad17956263f9cf70ba56b5