google/debug: ITMON:Enabling ITMON earlier in the PM resume cycle.
This patch changes in which phase the ITMON
driver resume is called:
originally the call was in the resume phase
while the PCIe controller resume was in the
resume_noirq phase and as a result the ITMON
did not enable to identify that the PCIe does
not respond (due to glitch in its clock which
triggered by modem crash).
Moving the ITMON to the resume_noirq pahse solved
the problem -- the ITMON intercept the fact that
the PCIe controller does not respond.

This issue is one of the use cases identified
due to the modem crash.
Bug: 318529894

Change-Id: I8fb18936b6d0fbf57efeae4362b7e1988a831d2e
Signed-off-by: Rami Gideoni <ramig@google.com>