usb: dwc3: prevent gadget binding during shutdown()
USB gadget will be stopped when receiving disconnect event. There
is a race between dwc3_otg_start_gadget() and userpsace re-binds
USB gadget. It may cause ITMON issues or other stability issues.
These issues were fixed by runtime PM and adding delay time.
However, it will increase the time for device reboot so that it
may trigger reboot timeout.

To fix this issue and race condition, acquire the dev lock of dwc3
gadget when USB cable was disconencted. The lock is to protect the
access to the driver of the device.

Bug: 325987370
Test: reboot for 7300+ times
Change-Id: I07d6da1b8fe9b129e8ece2f5b02fa8121570d0cd
Signed-off-by: Ray Chi <raychi@google.com>