drivers: thermals: s2mpg15: disable lpf before meter reset
If HW LPF is enabled, and meter control is reset dynamically
the HW uses the unfiltered stale data from before sensors are ready
(value of 0v=131C) right after the reset causing it to trigger a NTC
fault and device reset.

To fix the issue client drivers, in this case odpm driver needs to
disable LPF before issuing a reset and then reenable after sensors are
ready. In order for odpm driver to do so, spmic thermal driver exposes
an API to enable and disable NTC LPF

Bug: 322027909
Test: Build and boot on device. verify fixes bug by changing odpm
sampling frequency

Change-Id: I25cb4b82e0ce18462ce7694668932d2e13d785cb
Signed-off-by: Ramya Subramanian <rsubr@google.com>