PIXEL: ufs: prevent sw cmvp test in hw mode
If the fips module is built to support sw delivered keys mode,
do not register the verification test if running in hw mode. This
situation only occurs on device which are explicitly built to
support sw delivered keys, where the veriication is not needed
when in hw delivered keys mode.

Bug: 332726753
Merged-In: Ia7bef7be586a85f335c2d31c1756488611738511
Change-Id: Ia7bef7be586a85f335c2d31c1756488611738511
Signed-off-by: Konstantin Vyshetsky <vkon@google.com>