drivers: thermal: spmic: Change the thermal zone enable order
During the driver probe, after the thermistor hardware initialization
the thermal zones are registered and enabled. But the probe continues to
request interrupt handler and stats handle after enabling thermal zone.
This can lead to a race condition, where the irq handler be called and
the stats handler is invoked before initialization. This causes a crash
in stats update.

To avoid this, the probe order is changed to enable thermal zone after
all the resources are enabled. In this case after the interrupt and
stats handler are requested.

Bug: 329017255
Test: Booted and ensured all the thermistors are available

Change-Id: I3164ab8627a903de99770c9c337f7f29329bcd1d
Signed-off-by: Ram Chandrasekar <rchandrasekar@google.com>