vh: sched: Don't reduce iowait boost prematurely
When reducing the sched_latency_ns to 8ms, the iowait boost can be
negatively impacted as we can end up reducing it prematurely as tasks
can be context switched faster.

This is a known limitation in the current code that exist upstream where
there's an implicit assumptions there will be no cpufreq_update_util()
request soon after.

Bug: 269111781
Signed-off-by: Qais Yousef <qyousef@google.com>
Change-Id: Id88ae205cb7f1fb3fbb4e39ba7ed76a4cecfa584