drivers: thermal: spmic: Access hardware only when ready
After probe, the sensor hardware is configured and the driver checks to
see if the sensors are reading a reasonable value. Accessing the
temperature or changing the thresholds can cause undefined behavior.

To avoid the same, add a check to return EAGAIN, when the thermal zone
callbacks are called before the sensor is ready.

Bug: 329017255
Test: Booted and ensured all the thermistors are available

Change-Id: I1cfedbad3c191739ed016206ffc39c5c796043c1
Signed-off-by: Ram Chandrasekar <rchandrasekar@google.com>