drivers: odpm: cancel alarm only if work was already queued
Timer to start periodic refresh is started when work is queued in the
probe function. However there are several exit points in probe prior to
this that calls the _remove which ends up cancelling a non existent
timer leading to kernel panics.

In order to fix the issue, the alarm is only cancelled if a work queue
is detected. Additionally, since odpm driver is registered as a devm
resource it is not required to call iio_device_unregister() in odpm_remove().
It will unregister itself.

Bug: 330336444
Test: build and boot on device

Change-Id: Ib912be65c6546bd41489a3ed994199a0f516aaf9
Signed-off-by: Ramya Subramanian <rsubr@google.com>