vh: sched: Make rampup response time more uniform and coherent
When a task is bursty, the time it takes it to ramp-up can vary due to
utilization invariance which stretches the runtime of the task.

Cater for this by ramping-up util_est more coherently based on the task
runtime so that if it becomes suddenly busy, we'll increase util_est
more coherently.

We do so by adding the runtime and approximate_util_avg() directly on
context switch and tick.

We could restrict this to important tasks only in the future.

This change is gated by enabling auto_dvfs_headroom which will slow down
the ramp-up time for steady state situations. But with this change
bursty changes for tasks that start to run for longer will not suffer
a big delay and still ramp-up uniformly if they suddenly become always
running task.

Bug: 315748906
Signed-off-by: Qais Yousef <qyousef@google.com>
Change-Id: Id32550be16dc98d5dc93456d1c4e06f95eda940b