iommu: pcie: whi: Set correct pgtable base for PAGE_SIZE != 4K
The value for the page table base register was calculated under the
assumption that the PAGE_SIZE macro is defined as 4kb. This breaks in
configurations where PAGE_SIZE is set to 16kb or 64kb. Let's define a
new local macro named PT_BASE_SHIFT to use for this calculation. This
value is determined by the programming interface of the hardware and is
unrelated to the page size that the operating system is using.

Bug: 326481268
Bug: 318888819
Bug: 308663811
Test: Build and flash kernel with 4kb and 16kb page sizes
Change-Id: Id52e895ca9e3776d66e4df72e3dd219b3ef841a5
Signed-off-by: Juan Yescas <jyescas@google.com>