vh: sched: Keep track of our own prev_exec_sum_runtime
So that we can make __update_util_est_invariance() more accurate.
se->prev_exec_sum_runtime is updated when we pick_next_task(). So if
a task is long running we rely on updates happening on tick. But if
there are multiple tasks on the rq we can end up with less accurate
accounting for long running tasks.

We still keep the check to ensure delta_exec is not larger than
TICK_USEC as bolts and braces but it is not expected to hit this now.

Bug: 315748906
Signed-off-by: Qais Yousef <qyousef@google.com>
Change-Id: I14f8fa4724ecd39fbe7833fa2020d4acc7c8d696