iommu: samsung: Use PT_BASE_SHIFT to calculate pgtable offsets
The value for the page table base register was calculated under the
assumption that the PAGE_SIZE macro is defined as 4kb. This breaks in
configurations where PAGE_SIZE is set to 16kb or 64kb. Let's use the
PT_BASE_SHIFT instead.

The drivers fixed are:

- samsung-iommu-fault.c for gs101 and gs201.
- samsung-iommu-fault-v9.c for zuma and zumapro.

Bug: 331998310
Test: Build and flash kernel with 4kb and 16kb page sizes
Change-Id: Ia89096682726b51bc7ae34f433ddec1860bfcb24
Merged-In: Ia89096682726b51bc7ae34f433ddec1860bfcb24
Signed-off-by: Juan Yescas <jyescas@google.com>