PIXEL: ufs: fix fill_prdt registration logic
FIPS 140-3 requires the CMVP module to control access to encryption
services, which is done within the ufs-pixel-fips140.ko module. In the
event the UFS CMVP module is not built, the pixel_ufs_crypto_fill_prdt()
function should be registered instead. An issue arises when the UFS CMVP
module is built for sw delivered encryption keys, but the device
operates in hw delivered encryption key mode. Fix code readibility and
simplify the logic.

Bug: 333092313
Test: `atest vts_kernel_encryption_test` on:
	r4 sw key module with fips mode disabled
	r4 sw key module with fips mode enabled
	c10 hw key module

Change-Id: I258aac68a61b5f59d23e293da15456b6c21e6422
Signed-off-by: Konstantin Vyshetsky <vkon@google.com>