tcpci_max77759: BATOILO lvl=4.8A when plugged in
Due to tolerance, when device is plugged into USB, BATOILO level needs
to be reduced by 200mA.
This needs to be voted for both source and sink.
bcl_usb_votable's callback can take more than a msec to execute so
this is invoked from its own workqueue to not block the rest of the
state machine.

Bug: 334171387
Change-Id: I8be6a4156130c20f79abdb6299e9426983354e6b
Signed-off-by: George Lee <geolee@google.com>