vh: sched: Add new knobs to control rampup time for a adpf/groups
These knobs only make sense when auto dvfs headroom is enabled. It
controls the util_est() signal to help with coherent rampup time when
tasks become busy running (or run for extended period of time).

The new multiplier will control how fast this signal will rampup.
A value of 0 will disable this rampup feature. A value of 1 is the
default and will rampup with default PELT value. A value of 2 will
double the rampup speed (half the rampup time). A value of 3 will trip
the speed and 4 will quadriple it.

Higher values than 4 will be rejected as this is considered aggressive.
But if good reasons come up, we can relax this restriction.

Bug: 336278581
Signed-off-by: Qais Yousef <qyousef@google.com>
Change-Id: Iae172e9d1c56d13dbf803cbc4a9c50cc2396fd48