vh: sched: Better detect when we need to rampup util_est for a task
There's a hard problem of detecting when tasks are periodic and don't
need rampup faster vs tasks that are going through a transition phase
and need help to ensure they reach their 'settled' util_avg coherently
as they get busier.

Improve the logic by keeping track of current and last delta_exec and
util_avg so that we can detect when the task util_avg is rising - which
means it is getting busier and needs help to catch up faster with work.

This should help avoid ramping up tasks that their util avg has settled
or falling. And should help tasks that sleep for short period of time to
continue to rampup faster as they continue to get busier.

Ensure we reset util_est values for tasks with 0 rampup_multiplier. And
ensure we don't reset it for tasks that are ramping up.

Bug: 336169796
Signed-off-by: Qais Yousef <qyousef@google.com>
Change-Id: I83fa62547cb53aac06fff6a7b421948dbbd03cd9