drivers: thermal: Add uclamp cluster cooling device
Add a uclamp CPU cluster cooling device module. This module will register
one cooling device per cpufrequency domain, which loosely maps to a CPU
cluster. This module is similar to the cpufreq cooling device, where it
will have states mapping to each cpufreq level. When a mitigation
request is placed in this cooling device, the request will be sent to
the scheduler to be translated to an appropriate uclamp value.

Bug: 338150183
Test: Validated that the cooling device is able to vote for uclamp_max

Change-Id: I901c1af5fce3f94fd5f5f38c2b35b3544fab4f44
Signed-off-by: Ram Chandrasekar <rchandrasekar@google.com>