vh: sched: Add min_load_cpu as another default target
When all allowed cpus of a task are fully busy such that there is
no cpu with spare capacity > 0 and all are group_overutilize, we
will keep the last cpu as most_spare_cap_cpu. If there are no
other candidates, it will use the most_spare_cap_cpu as default
target. This may cause many tasks wake up on the same cpu.

To solve it, we add min_load_cpu as another default target if there
is no cpu with spare capacity > 0.

Bug: 335875684
Change-Id: I96d104ea6867cef9599b20e54c5c3e9d2b741d0e
Signed-off-by: Rick Yiu <rickyiu@google.com>