arm64/s2mpu: Add deny-all S2MPUs
Some S2MPUs are not currently in the device tree on the assumption
that their initiator only issues secure traffic.
However, SysMMU has an emulation feature that can be misused to read
from arbitrary memory locations, and with SysMMU under the control of
the kernel, we need to configure S2MPU to block such potentially
malicious transactions.

For that we add a new flag deny-all to S2MPUs which does the following:
- The S2MPU will be put in deny all state at boot.
- The S2MPU doesn’t get power managed.
- The S2MPU doesn’t idmap the kernel page table.

Bug: 204991862
Bug: 342092765
Change-Id: Ifbf6e94a683e5c6d80098666ad92d142feb246b1
Signed-off-by: Mostafa Saleh <smostafa@google.com>